+++ eras/conway/impl/src/Cardano/Ledger/Conway/Governance/Procedures.hs
@@ -82,11 +82,14 @@ import Cardano.Slotting.Slot (EpochNo)
 import Control.DeepSeq (NFData (..))
 import Control.Monad (when)
 import Data.Aeson (
+  FromJSON (..),
   KeyValue (..),
   ToJSON (..),
   ToJSONKey (..),
   object,
   pairs,
+  withObject,
+  (.:),
  )
 import Data.Aeson.Types (toJSONKeyText)
 import Data.Map.Strict (Map)
@@ -94,6 +97,7 @@ import Data.Maybe.Strict (StrictMaybe (..))
 import Data.Sequence (Seq (..))
 import Data.Set (Set)
 import qualified Data.Text as Text
+import Data.Unit.Strict (forceElemsToWHNF)
 import Data.Word (Word32)
 import GHC.Generics (Generic)
 import Lens.Micro (Lens', lens)
@@ -373,6 +377,14 @@ instance Era era => EncCBOR (Committee era) where
 instance EraPParams era => ToJSON (Committee era) where
   toJSON = object . toCommitteePairs
   toEncoding = pairs . mconcat . toCommitteePairs
+
+instance Era era => FromJSON (Committee era) where
+  parseJSON = withObject "Committee" parseCommittee
+    where
+      parseCommittee o =
+        Committee
+          <$> (forceElemsToWHNF <$> o .: "members")
+          <*> o .: "quorum"
 
 toCommitteePairs :: (KeyValue a, EraPParams era) => Committee era -> [a]
 toCommitteePairs committee@(Committee _ _) =